Refactor MetaSwap to inherit Swap
