+++ contracts/Swap.sol
@@ -204,7 +204,7 @@ contract Swap is OwnerPausableUpgradeable, ReentrancyGuardUpgradeable {
      * @dev See the StableSwap paper for details
      * @return A parameter
      */
-    function getA() external view returns (uint256) {
+    function getA() external view virtual returns (uint256) {
         return swapStorage.getA();
     }
 
@@ -213,7 +213,7 @@ contract Swap is OwnerPausableUpgradeable, ReentrancyGuardUpgradeable {
      * @dev See the StableSwap paper for details
      * @return A parameter in its raw precision form
      */
-    function getAPrecise() external view returns (uint256) {
+    function getAPrecise() external view virtual returns (uint256) {
         return swapStorage.getAPrecise();
     }
 
@@ -222,7 +222,7 @@ contract Swap is OwnerPausableUpgradeable, ReentrancyGuardUpgradeable {
      * @param index the index of the token
      * @return address of the token at given index
      */
-    function getToken(uint8 index) public view returns (IERC20) {
+    function getToken(uint8 index) public view virtual returns (IERC20) {
         require(index < swapStorage.pooledTokens.length, "Out of range");
         return swapStorage.pooledTokens[index];
     }
@@ -233,7 +233,12 @@ contract Swap is OwnerPausableUpgradeable, ReentrancyGuardUpgradeable {
      * @param tokenAddress address of the token
      * @return the index of the given token address
      */
-    function getTokenIndex(address tokenAddress) public view returns (uint8) {
+    function getTokenIndex(address tokenAddress)
+        public
+        view
+        virtual
+        returns (uint8)
+    {
         uint8 index = tokenIndexes[tokenAddress];
         require(
             address(getToken(index)) == tokenAddress,
@@ -246,7 +251,12 @@ contract Swap is OwnerPausableUpgradeable, ReentrancyGuardUpgradeable {
      * @notice Return timestamp of last deposit of given address
      * @return timestamp of the last deposit made by the given address
      */
-    function getDepositTimestamp(address user) external view returns (uint256) {
+    function getDepositTimestamp(address user)
+        external
+        view
+        virtual
+        returns (uint256)
+    {
         return swapStorage.getDepositTimestamp(user);
     }
 
@@ -255,7 +265,12 @@ contract Swap is OwnerPausableUpgradeable, ReentrancyGuardUpgradeable {
      * @param index the index of the token
      * @return current balance of the pooled token at given index with token's native precision
      */
-    function getTokenBalance(uint8 index) external view returns (uint256) {
+    function getTokenBalance(uint8 index)
+        external
+        view
+        virtual
+        returns (uint256)
+    {
         require(index < swapStorage.pooledTokens.length, "Index out of range");
         return swapStorage.balances[index];
     }
@@ -264,7 +279,7 @@ contract Swap is OwnerPausableUpgradeable, ReentrancyGuardUpgradeable {
      * @notice Get the virtual price, to help calculate profit
      * @return the virtual price, scaled to the POOL_PRECISION_DECIMALS
      */
-    function getVirtualPrice() external view returns (uint256) {
+    function getVirtualPrice() external view virtual returns (uint256) {
         return swapStorage.getVirtualPrice();
     }
 
@@ -280,7 +295,7 @@ contract Swap is OwnerPausableUpgradeable, ReentrancyGuardUpgradeable {
         uint8 tokenIndexFrom,
         uint8 tokenIndexTo,
         uint256 dx
-    ) external view returns (uint256) {
+    ) external view virtual returns (uint256) {
         return swapStorage.calculateSwap(tokenIndexFrom, tokenIndexTo, dx);
     }
 
@@ -304,7 +319,7 @@ contract Swap is OwnerPausableUpgradeable, ReentrancyGuardUpgradeable {
         address account,
         uint256[] calldata amounts,
         bool deposit
-    ) external view returns (uint256) {
+    ) external view virtual returns (uint256) {
         return swapStorage.calculateTokenAmount(account, amounts, deposit);
     }
 
@@ -318,6 +333,7 @@ contract Swap is OwnerPausableUpgradeable, ReentrancyGuardUpgradeable {
     function calculateRemoveLiquidity(address account, uint256 amount)
         external
         view
+        virtual
         returns (uint256[] memory)
     {
         return swapStorage.calculateRemoveLiquidity(account, amount);
@@ -336,7 +352,7 @@ contract Swap is OwnerPausableUpgradeable, ReentrancyGuardUpgradeable {
         address account,
         uint256 tokenAmount,
         uint8 tokenIndex
-    ) external view returns (uint256 availableTokenAmount) {
+    ) external view virtual returns (uint256 availableTokenAmount) {
         return
             swapStorage.calculateWithdrawOneToken(
                 account,
@@ -357,6 +373,7 @@ contract Swap is OwnerPausableUpgradeable, ReentrancyGuardUpgradeable {
     function calculateCurrentWithdrawFee(address user)
         external
         view
+        virtual
         returns (uint256)
     {
         return swapStorage.calculateCurrentWithdrawFee(user);
@@ -367,7 +384,12 @@ contract Swap is OwnerPausableUpgradeable, ReentrancyGuardUpgradeable {
      * @param index Index of the pooled token
      * @return admin's token balance in the token's precision
      */
-    function getAdminBalance(uint256 index) external view returns (uint256) {
+    function getAdminBalance(uint256 index)
+        external
+        view
+        virtual
+        returns (uint256)
+    {
         return swapStorage.getAdminBalance(index);
     }
 
@@ -389,6 +411,7 @@ contract Swap is OwnerPausableUpgradeable, ReentrancyGuardUpgradeable {
         uint256 deadline
     )
         external
+        virtual
         nonReentrant
         whenNotPaused
         deadlineCheck(deadline)
@@ -411,6 +434,7 @@ contract Swap is OwnerPausableUpgradeable, ReentrancyGuardUpgradeable {
         uint256 deadline
     )
         external
+        virtual
         nonReentrant
         whenNotPaused
         deadlineCheck(deadline)
@@ -433,7 +457,13 @@ contract Swap is OwnerPausableUpgradeable, ReentrancyGuardUpgradeable {
         uint256 amount,
         uint256[] calldata minAmounts,
         uint256 deadline
-    ) external nonReentrant deadlineCheck(deadline) returns (uint256[] memory) {
+    )
+        external
+        virtual
+        nonReentrant
+        deadlineCheck(deadline)
+        returns (uint256[] memory)
+    {
         return swapStorage.removeLiquidity(amount, minAmounts);
     }
 
@@ -453,6 +483,7 @@ contract Swap is OwnerPausableUpgradeable, ReentrancyGuardUpgradeable {
         uint256 deadline
     )
         external
+        virtual
         nonReentrant
         whenNotPaused
         deadlineCheck(deadline)
@@ -482,6 +513,7 @@ contract Swap is OwnerPausableUpgradeable, ReentrancyGuardUpgradeable {
         uint256 deadline
     )
         external
+        virtual
         nonReentrant
         whenNotPaused
         deadlineCheck(deadline)
