--- contracts/meta/AmplificationUtils.sol
@@ -1,94 +1,0 @@
-import "@openzeppelin/contracts/token/ERC20/SafeERC20.sol";
-import "./MetaSwapUtils.sol";
-
-pragma solidity 0.6.12;
-
-library AmplificationUtils {
-    using SafeMath for uint256;
-
-    event RampA(
-        uint256 oldA,
-        uint256 newA,
-        uint256 initialTime,
-        uint256 futureTime
-    );
-    event StopRampA(uint256 currentA, uint256 time);
-
-    // Constant values used in ramping A calculations
-    uint256 public constant A_PRECISION = 100;
-    uint256 public constant MAX_A = 10**6;
-    uint256 private constant MAX_A_CHANGE = 2;
-    uint256 private constant MIN_RAMP_TIME = 14 days;
-
-    /**
-     * @notice Start ramping up or down A parameter towards given futureA_ and futureTime_
-     * Checks if the change is too rapid, and commits the new A value only when it falls under
-     * the limit range.
-     * @param self Swap struct to update
-     * @param futureA_ the new A to ramp towards
-     * @param futureTime_ timestamp when the new A should be reached
-     */
-    function rampA(
-        MetaSwapUtils.Swap storage self,
-        uint256 initialAPrecise,
-        uint256 futureA_,
-        uint256 futureTime_
-    ) external {
-        require(
-            block.timestamp >= self.initialATime.add(1 days),
-            "Wait 1 day before starting ramp"
-        );
-        require(
-            futureTime_ >= block.timestamp.add(MIN_RAMP_TIME),
-            "Insufficient ramp time"
-        );
-        require(
-            futureA_ > 0 && futureA_ < MAX_A,
-            "futureA_ must be > 0 and < MAX_A"
-        );
-
-        uint256 futureAPrecise = futureA_.mul(A_PRECISION);
-
-        if (futureAPrecise < initialAPrecise) {
-            require(
-                futureAPrecise.mul(MAX_A_CHANGE) >= initialAPrecise,
-                "futureA_ is too small"
-            );
-        } else {
-            require(
-                futureAPrecise <= initialAPrecise.mul(MAX_A_CHANGE),
-                "futureA_ is too large"
-            );
-        }
-
-        self.initialA = initialAPrecise;
-        self.futureA = futureAPrecise;
-        self.initialATime = block.timestamp;
-        self.futureATime = futureTime_;
-
-        emit RampA(
-            initialAPrecise,
-            futureAPrecise,
-            block.timestamp,
-            futureTime_
-        );
-    }
-
-    /**
-     * @notice Stops ramping A immediately. Once this function is called, rampA()
-     * cannot be called for another 24 hours
-     * @param self Swap struct to update
-     */
-    function stopRampA(MetaSwapUtils.Swap storage self, uint256 currentA)
-        external
-    {
-        require(self.futureATime > block.timestamp, "Ramp is already stopped");
-
-        self.initialA = currentA;
-        self.futureA = currentA;
-        self.initialATime = block.timestamp;
-        self.futureATime = block.timestamp;
-
-        emit StopRampA(currentA, block.timestamp);
-    }
-}
