+++ deploy/004_deploy_MetaSwapUtils.ts
@@ -1,0 +1,20 @@
+import { HardhatRuntimeEnvironment } from "hardhat/types"
+import { DeployFunction } from "hardhat-deploy/types"
+
+const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
+  const { deployments, getNamedAccounts } = hre
+  const { deploy, get } = deployments
+  const { deployer } = await getNamedAccounts()
+
+  await deploy("MetaSwapUtils", {
+    from: deployer,
+    log: true,
+    libraries: {
+      MathUtils: (await get("MathUtils")).address,
+    },
+    skipIfAlreadyDeployed: true,
+  })
+}
+export default func
+func.tags = ["MetaSwapUtils"]
+func.dependencies = ["MathUtils"]
